Fix a few incorrect malloc declarations in ntdrivers and ssh directories

+++ c/ssh/s3_srvr.blast.13_false-unreach-call.i.cil.c
@@ -1,6 +1,6 @@
 extern void __VERIFIER_error() __attribute__ ((__noreturn__));
 
-extern void *malloc(unsigned long sz );
+extern void *malloc(unsigned int sz );
 extern char __VERIFIER_nondet_char(void);
 extern int __VERIFIER_nondet_int(void);
 extern long __VERIFIER_nondet_long(void);
